Revert "list.m4: Add some new pthread symbols"
This reverts commit fa6db2f9c018c54a47e94c0175450303d700aa92.

This commit causes pthread-stubs to provide a library (with stubs) for
most platforms. This by itself is the goal of the project, although
unveiled severe design issue with the project as a whole.

That will be covered and addressed in a later commit.

In other words: pthread-stubs won't provide/guarantee these symbols any
more. That is perfectly fine since a) there has been no pthread-stubs
release that provides the API, and b) these should be provided by a full
blown pthreads implementation.

Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=98048

Acked-by: Eric Anholt <eric@anholt.net>
Signed-off-by: Uli Schlachter <psychon@znc.in>